refactor(sysdep): вынести сетевые системные заголовки в sysdep_net.h (#3228)
В sysdep.h было два больших блока, защищённых #if defined(__COMM_C__):

* системные заголовки: <sys/socket.h>, <netinet/in.h>, <arpa/inet.h>,
  <netdb.h>, <signal.h>, <sys/select.h>, <sys/wait.h>, <sys/uio.h>,
  <sys/resource.h>, <fcntl.h>, struct in_addr fallback;
* прототипы сетевых функций: accept/bind/listen/socket/htonl/inet_addr
  и т.п.

Макрос __COMM_C__ определялся только в comm.cpp перед include "comm.h",
после чего sysdep.h раскрывал эти блоки только для одного TU. Внешне
это выглядело как "sysdep.h тащит всё", по факту -- ничего, кроме
комм-стека для comm.cpp.

Проблемы такого подхода:

* непрозрачно: чтобы понять, какой набор заголовков получит TU, нужно
  знать про скрытый макрос;
* блокирует sysdep.h в precompiled headers (PCH компилируется без
  TU-локального #define, поэтому эти блоки в PCH теряются и comm.cpp
  потом не находит sockaddr_in).

Изменения:

* sysdep_net.h -- новый заголовок со всеми сетевыми системными
  include'ами и прототипами.
* sysdep.h -- из него удалены оба __COMM_C__-блока, оставлено общее
  системное ядро (stdio/stdlib/errno/time/zlib/limits/...).
* comm.cpp -- заменил "#define __COMM_C__" на явный
  "#include \"engine/core/sysdep_net.h\"".

Никакие другие файлы трогать не пришлось: остальные потребители
сокетов (iosystem.cpp, interpreter.cpp, admin_api.cpp, char_player.cpp)
и так инклудят системные заголовки явно.
